requirements.txt
alabaster==0.7.12
asgiref==3.5.2
Babel==2.10.3
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
Django==4.0.4
djangorestframework==3.13.1
docutils==0.17.1
factory-boy==3.2.1
Faker==13.12.0
gunicorn==20.1.0
h11==0.13.0
idna==3.3
imagesize==1.3.0
Jinja2==3.1.2
Mako==1.2.0
Markdown==3.3.7
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.1.1
psycopg2==2.9.3
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-monkey-business==1.0.0
pytz==2022.1
requests==2.28.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx==5.0.2
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
urllib3==1.26.9
uvicorn==0.17.6
